third_party/WebKit/LayoutTests/printing/css2.1/page-break-inside-000-expected.txt

There must be a page break below between "FIRST" and "LAST"
FIRSTdummy
        dummy
        dummy
        dummy
        dummy
        dummy
        dummy
        dummy
        dummyFIRST


PASS: page number of "test1" is 0
PASS: page number of "test2" is 1
All tests passed
